+++ docs/changelog.rst
@@ -7,13 +7,13 @@ Changelog
 .. note::
    Under development
 
-- Introduce :mod:`gidgethub.app`.
+- Introduce :mod:`gidgethub.apps`.
 
-- Introduce :meth:`gidgethub.app.get_installation_access_token`
+- Introduce :meth:`gidgethub.apps.get_installation_access_token`
   for obtaining an installation access token that can be used to authenticate as
   a GitHub App installation. (`Issue #71 <https://github.com/brettcannon/gidgethub/issues/71>`_).
 
-- Introduce :meth:`gidgethub.app.get_jwt` for constructing the JSON Web Token
+- Introduce :meth:`gidgethub.apps.get_jwt` for constructing the JSON Web Token
   that can be used to access endpoints that require it.
 
 4.0.0